app/requirements.txt
Flask==3.1.0
passlib==1.7.4
markdown==3.7
bleach==6.2.0
cryptography==44.0.0
pycryptodome==3.21.0
requests==2.32.3
pyotp==2.9.0
qrcode==8.0
Flask-WTF==1.2.2
Pillow==11.1.0
uwsgi==2.0.28
python-dotenv==1.0.1